+++ CHANGES.md
@@ -4,6 +4,7 @@ Zing Changelog
 v.next (not released yet)
 -------------------------
 
+* Removed captcha middleware and associated setting.
 * Removed `ZING_INSTANCE_ID`. This can easily be implemented by anyone.
 * Cleaned up setting names and removed obsoletes:
   * `POOTLE_TM_SERVER` -> `ZING_TM_SERVER`